Split heavy /api routes out of api-app.ts into src/api/routes/
api-app.ts had grown to 427 lines, with three POST routes (compare /
compare-renderers / reason) each carrying its own playwright + heatmap
+ pipeline imports. The assembler-vs-handler split is the same shape
already used under src/cli/workflow/.

- src/api/routes/compare.ts — /api/compare (HTML → multi-viewport
  pixel + computed-style diff).
- src/api/routes/compare-renderers.ts — /api/compare-renderers
  (Chromium vs Crater BiDi cross-engine diff); takes its
  resolveCraterAvailable hook through registerCompareRenderersRoute.
- src/api/routes/reason.ts — /api/reason (VLM analyze + LLM fix
  suggestions via vrt-ai's reasoning pipeline).
- src/api/routes/helpers.ts — shared resolveHtmlSource (with SSRF
  guard) and loadCraterAvailability.

api-app.ts is now a 97-line assembler that holds the body-size
middleware, the two GET endpoints (/api/openapi.json, /api/status),
the small /api/smoke-test POST, and calls each register*Route helper.
No behavior change; routes were copied verbatim apart from rewriting
the css-challenge-core import path for the new file depth.
